Extract the shared cd=26 network-info registration
Venus and Jupiter answer `cd=26` with the same payload and registered the
same five sensors from two near-identical copies of the field loop. Move it
into `device/networkInfoBase.ts`, following the existing `meterBase.ts`
pattern, with the command code, the message predicate and whether the IP
sensor is enabled by default as the only per-device knobs.

`VenusNetworkInfo` and `JupiterNetworkInfo` become aliases of the shared
`NetworkInfo` shape. No behaviour change: Venus still enables the IP Address
sensor by default and Jupiter still enables none, and every sensor still
waits for the device to report its value.
